refactor: #6 move Json response related logic from application controller to concern
